backend/contact_form/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-secret-manager==2.*
sendgrid==6.*
